legacy/requirements.txt
tonutils
tonsdk
aio_pika
python-dotenv
typing_extensions
fastapi
uvicorn
prometheus_client
